+++ src/Compilers/CSharp/Portable/Symbols/Source/SourceNamedTypeSymbol.cs
@@ -19,46 +19,6 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.Symbols
 {
-    internal sealed class TypeParameterInfo
-    {
-        public ImmutableArray<TypeParameterSymbol> LazyTypeParameters;
-
-        /// <summary>
-        /// A collection of type parameter constraint types, populated when
-        /// constraint types for the first type parameter are requested.
-        /// </summary>
-        public ImmutableArray<ImmutableArray<TypeWithAnnotations>> LazyTypeParameterConstraintTypes;
-
-        /// <summary>
-        /// A collection of type parameter constraint kinds, populated when
-        /// constraint kinds for the first type parameter are requested.
-        /// </summary>
-        public ImmutableArray<TypeParameterConstraintKind> LazyTypeParameterConstraintKinds;
-
-        public static readonly TypeParameterInfo Empty = new TypeParameterInfo(
-            ImmutableArray<TypeParameterSymbol>.Empty, ImmutableArray<ImmutableArray<TypeWithAnnotations>>.Empty, ImmutableArray<TypeParameterConstraintKind>.Empty);
-
-        public TypeParameterInfo() : this(default, default, default)
-        {
-        }
-
-        private TypeParameterInfo(
-            ImmutableArray<TypeParameterSymbol> typeParameters,
-            ImmutableArray<ImmutableArray<TypeWithAnnotations>> typeParameterConstraintTypes,
-            ImmutableArray<TypeParameterConstraintKind> typeParameterConstraintKinds)
-        {
-            LazyTypeParameters = typeParameters;
-            LazyTypeParameterConstraintTypes = typeParameterConstraintTypes;
-            LazyTypeParameterConstraintKinds = typeParameterConstraintKinds;
-        }
-
-        public static TypeParameterInfo Create(ImmutableArray<TypeParameterSymbol> typeParameters)
-        {
-            // If we have no type parameters (common case), we can just point at the singleton empty instance.
-            return typeParameters.IsEmpty ? Empty : new TypeParameterInfo(typeParameters, default, default);
-        }
-    }
-
     // This is a type symbol associated with a type definition in source code.
     // That is, for a generic type C<T> this is the instance type C<T>.  
     internal sealed partial class SourceNamedTypeSymbol : SourceMemberContainerTypeSymbol, IAttributeTargetSymbol
